Replace `fsm_trim()` with `fsm_collect_unreachable_states()`.
This does the same thing, but better.
