+++ src/libfsm/Makefile
@@ -1,7 +1,6 @@
 .include "../../share/mk/top.mk"
 
 SRC += src/libfsm/collate.c
-SRC += src/libfsm/collect.c
 SRC += src/libfsm/complete.c
 SRC += src/libfsm/clone.c
 SRC += src/libfsm/eclosure.c
