+++ src/libre/re_comp.c
@@ -111,7 +111,7 @@ re_comp_ast(struct ast_re *ast,
 
 	if (!comp_iter(&env, x, y, ast->expr)) { goto error; }
 
-	if (-1 == fsm_collect_unreachable_states(env.fsm)) { goto error; }
+	if (-1 == fsm_trim(env.fsm)) { goto error; }
 
 	return env.fsm;
